requirements.txt
# Core LangChain dependencies
langchain>=0.1.0
langchain-core>=0.1.0
langchain-community>=0.0.20
langchain-google-genai>=0.0.6
langchain-huggingface>=0.0.6
langchain-openai>=0.0.6
langchain-ollama>=0.0.1

# Vector store and embeddings
faiss-cpu>=1.7.4
sentence-transformers>=2.2.2

# Web framework
fastapi>=0.104.0
uvicorn>=0.24.0

# Discord integration
discord.py>=2.3.0

# Environment and utilities
python-dotenv>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0

# Security
pynacl>=1.5.0

# HTTP requests
requests>=2.31.0

# Scientific computing
numpy>=1.24.0
